+++ lib/models/stream_signal.dart
@@ -1,23 +1,25 @@
 part of 'signal.dart';
 
-class StreamSignal<T> extends Signal<AsyncValue<T>> {
+class StreamSignal<T> extends Signal<T?> {
   final Stream<T> _source;
   StreamSubscription<T>? _subscription;
 
-  StreamSignal(this._source) : super(AsyncData.initial()) {
+  StreamSignal(this._source) : super(null) {
     _listen();
   }
 
+  T? get rawValue => value;
+
   void _listen() {
     _subscription?.cancel();
-    _emit(const AsyncLoading());
+    _emit(null);
     _subscription = _source.listen(
-      (data) => _emit(AsyncData(data)),
-      onError: (e, StackTrace st) => _emit(AsyncError(e, st)),
+      (data) => _emit(data),
+      onError: (e, StackTrace st) => _emit(null),
     );
   }
 
-  void _emit(AsyncValue<T> newValue) {
+  void _emit(T? newValue) {
     value = newValue;
     controller.add(newValue);
   }